feat: Add stdio adapter, examples folder, and complete MCP integration
- Added ine_mcp.stdio module for direct MCP client connections (Cursor/Claude)
- Fixed timestamp parsing bug in aggregation.py (int64 cast)
- Created comprehensive examples/ folder with 6 examples + outputs
- Added simple_search.py, ipc_analysis.py for beginners
- Added tourism/housing correlation analysis (national and provincial attempts)
- Added investigate_ine_operations.py for API exploration
- Documented INE API limitations (provincial data not in Tempus API)
- Organized all analysis scripts and outputs in examples/
- Updated examples/README.md with learning path and lessons learned
- Removed redundant mcp_config_snippet.json from root

+++ ine_mcp_python/ine_mcp/tools/aggregation.py
@@ -104,7 +104,7 @@ class DataAggregator:
         # INE API returns dates like 1706742000000 for some series
         numeric_mask = pd.to_numeric(date_series, errors='coerce').notna()
         if numeric_mask.any():
-            timestamps = pd.to_numeric(date_series[numeric_mask], errors='coerce')
+            timestamps = pd.to_numeric(date_series[numeric_mask], errors='coerce').astype(np.int64)
             # Convert milliseconds to datetime
             result[numeric_mask] = pd.to_datetime(timestamps, unit='ms', errors='coerce')
